+++ extract.hpp
@@ -11,10 +11,10 @@ class CmdExtract
 
     ComprDataIO DataIO;
     Unpack *Unp;
-    long TotalFileCount;
+    unsigned long TotalFileCount;
 
-    long FileCount;
-    long MatchedArgs;
+    unsigned long FileCount;
+    unsigned long MatchedArgs;
     bool FirstFile;
     bool AllMatchesExact;
     bool ReconstructDone;
@@ -33,9 +33,9 @@ class CmdExtract
     ~CmdExtract();
     void DoExtract(CommandData *Cmd);
     void ExtractArchiveInit(CommandData *Cmd,Archive &Arc);
-    bool ExtractCurrentFile(CommandData *Cmd,Archive &Arc,int HeaderSize,
+    bool ExtractCurrentFile(CommandData *Cmd,Archive &Arc,size_t HeaderSize,
                             bool &Repeat);
-    static void UnstoreFile(ComprDataIO &DataIO,Int64 DestUnpSize);
+    static void UnstoreFile(ComprDataIO &DataIO,int64 DestUnpSize);
 
     bool SignatureFound;
 };
